+++ nnmnkwii/preprocessing/__init__.py
@@ -3,12 +3,14 @@ from __future__ import division, print_function, absolute_import
 
 from .generic import preemphasis, inv_preemphasis
 from .generic import delta_features, trim_zeros_frames, remove_zeros_frames
-from .generic import adjast_frame_length, scale, minmax_scale
+from .generic import adjast_frame_length, adjast_frame_lengths
+from .generic import scale, minmax_scale
 from .generic import meanvar, meanstd, minmax
 from .f0 import interp1d
 from .modspec import modspec, modphase
 
 __all__ = ['preemphasis', 'inv_preemphasis', 'delta_features',
            'trim_zeros_frames', 'remove_zeros_frames',
-           'adjast_frame_length', 'scale', 'minmax_scale', 'meanvar',
+           'adjast_frame_length', 'adjast_frame_lengths',
+           'scale', 'minmax_scale', 'meanvar',
            'meanstd', 'minmax', 'interp1d', 'modspec', 'modphase']