+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Change log
 v0.0.6 <2017-xx-xx>
 ------------------
 
+- `#34`_: Add :func:`nnmnkwii.preprocessing.adjast_frame_length`.
+- `#34`_: ``adjast_frame_lengths`` now supports ``divisible_by`` parameter. ``ensure_even`` is deprecated.
+- `#34`_: Rename ``adjast_frame_length`` to ``adjast_Frame_lengths``
 - Add references to :func:`nnmnkwii.postfilters.merlin_post_filter`.
 
 v0.0.5 <2017-09-19>
@@ -54,3 +57,4 @@ v0.0.1 <2017-08-14>
 .. _#27: https://github.com/r9y9/nnmnkwii/pull/27
 .. _#28: https://github.com/r9y9/nnmnkwii/pull/28
 .. _#31: https://github.com/r9y9/nnmnkwii/pull/31
+.. _#34: https://github.com/r9y9/nnmnkwii/pull/34